4366: move practitioner form into separate component for reuse
